requirements.txt
azure-cognitiveservices-speech==1.24.2
monpa==0.3.3
fastapi==0.89.1
uvicorn==0.20.0
python-docx==0.8.11
python-multipart==0.0.5
requests-toolbelt==0.10.1
scipy==1.10.0
gensim==4.3.0
numpy==1.23.5
pandas==1.5.1
protobuf==3.19.6
requests==2.28.1
srt==3.5.2
scikit-learn==1.1.2
anyio==3.6.2
SoundFile==0.10.3.post1
networkx==2.8.5
librosa==0.9.2
ipython==8.7.0
ipykernel==6.19.2
ipython-genutils==0.2.0
audioread==3.0.0